moving tests to match prosper.datareader.news framework
