+++ tests/test_stocks.py
@@ -1,3 +1,4 @@
+"""test prosper.datareader.stocks top-level functions"""
 from os import path
 
 import pytest
@@ -52,148 +53,3 @@ class TestGetQuoteRH:
             'updated_at', 'url', 'volume', 'year_founded', 'change_pct'
         ]
         assert set(list(no_filter.columns.values)) == set(all_keys)
-
-class TestCompanyNewsRobinhood:
-    """valdiate stocks.company_news_rh"""
-    good_ticker = helpers.CONFIG.get('STOCKS', 'good_ticker')
-    bad_ticker = helpers.CONFIG.get('STOCKS', 'bad_ticker')
-    expected_news_cols = [
-        'api_source', 'author', 'instrument', 'num_clicks', 'preview_image_url',
-        'published_at', 'relay_url', 'source', 'summary', 'title', 'updated_at',
-        'url', 'uuid', 'preview_image_width', 'preview_image_height'
-    ]
-
-    def test_company_news_rh_happypath(self):
-        """make sure production endpoint works as expected"""
-        all_news_df = stocks.company_news_rh(self.good_ticker)
-
-        assert isinstance(all_news_df, pd.DataFrame)
-
-        unique_values, unique_expected = helpers.find_uniques(
-            list(all_news_df.columns.values),
-            self.expected_news_cols
-        )
-        assert unique_expected == []
-        if unique_values:
-            pytest.xfail(
-                'Unexpected values from company_news_rh(): {}'.format(unique_values)
-            )
-
-    @pytest.mark.long
-    def test_vader_application(self):
-        """make sure use-case for news + vader works"""
-        import prosper.datareader.utils as utils
-        all_news_df = stocks.company_news_rh(self.good_ticker)
-
-        graded_news = utils.vader_sentiment(all_news_df, 'title')
-
-        expected_cols = self.expected_news_cols
-        expected_cols.extend(['neu', 'pos', 'compound', 'neg'])
-
-        unique_values, unique_expected = helpers.find_uniques(
-            list(graded_news.columns.values),
-            expected_cols
-        )
-        assert unique_expected == []
-        if unique_values:
-            pytest.xfail(
-                'Unexpected values from vader_sentiment(): {}'.format(unique_values)
-            )
-
-class TestMarketNewsGoogle:
-    """valdiate behavior on stocks.market_news_google()"""
-    @flaky
-    def test_market_news_google_happypath(self):
-        """validate default behavior"""
-        all_news_df = stocks.market_news_google()
-
-        assert isinstance(all_news_df, pd.DataFrame)
-
-        expected_cols = [
-            'age', 'primary', 'source', 'blurb',
-            'sru', 'title', 'datetime', 'url', 'usg'
-        ]
-
-        assert list(all_news_df.columns.values) == expected_cols
-
-    @flaky
-    def test_market_news_google_no_pretty(self):
-        """validate not-pretty return works as expected"""
-        all_news_df = stocks.market_news_google(pretty=False)
-
-        assert isinstance(all_news_df, pd.DataFrame)
-
-        expected_cols = [
-            'd', 'primary', 's', 'sp', 'sru', 't', 'tt', 'u', 'usg'
-        ]
-
-        assert list(all_news_df.columns.values) == expected_cols
-
-
-class TestCompanyNewsGoogle:
-    """validate behavior for stocks.company_news_google()"""
-    good_ticker = helpers.CONFIG.get('STOCKS', 'good_ticker')
-    bad_ticker = helpers.CONFIG.get('STOCKS', 'bad_ticker')
-
-    @flaky
-    def test_company_news_happypath(self):
-        """make sure pandas wrapper works as expected"""
-        all_news_df = stocks.company_news_google(self.good_ticker)
-
-        assert isinstance(all_news_df, pd.DataFrame)
-
-        expected_cols = [
-            'age', 'primary', 'source', 'blurb',
-            'sru', 'title', 'datetime', 'url', 'usg'
-        ]
-
-        unique_values, unique_expected = helpers.find_uniques(
-            list(all_news_df.columns.values),
-            expected_cols
-        )
-        assert unique_expected == []
-        if unique_values:
-            pytest.xfail('Unexpected values from company_news_google(): {}'.format(unique_values))
-
-    @flaky
-    def test_company_news_no_pretty(self):
-        """validate not-pretty return works as expected"""
-        all_news_df = stocks.company_news_google(
-            self.good_ticker,
-            pretty=False
-        )
-
-        assert isinstance(all_news_df, pd.DataFrame)
-
-        expected_cols = [
-            'd', 'primary', 's', 'sp', 'sru', 't', 'tt', 'u', 'usg'
-        ]
-        unique_values, unique_expected = helpers.find_uniques(
-            list(all_news_df.columns.values),
-            expected_cols
-        )
-        assert unique_expected == []
-        if unique_values:
-            pytest.xfail('Unexpected values from company_news_google(): {}'.format(unique_values))
-
-    @flaky
-    def test_vader_application(self):
-        """make sure use-case for news + vader works"""
-        import prosper.datareader.utils as utils
-        all_news_df = stocks.company_news_google(self.good_ticker)
-
-        graded_news = utils.vader_sentiment(all_news_df, 'title')
-
-        expected_cols = [
-            'age', 'primary', 'source', 'blurb',
-            'sru', 'title', 'datetime', 'url', 'usg',
-            'neu', 'pos', 'compound', 'neg'
-        ]
-
-        unique_values, unique_expected = helpers.find_uniques(
-            list(graded_news.columns.values),
-            expected_cols
-        )
-        assert unique_expected == []
-        if unique_values:
-            pytest.xfail('Unexpected values from company_news_google(): {}'.format(unique_values))
